# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rental now Api Documentation",
        "termsOfService": "http://www.rentalnow.com.au/",
        "contact": {
            "name": "Jagdish kunwar",
            "email": "jagdishkunwar002@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/houses/{house_id}/images": {
            "post": {
                "tags": [
                    "Houses"
                ],
                "summary": "upload house images",
                "description": "upload house images",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\House\\HouseController::storeImages",
                "parameters": [
                    {
                        "name": "house_id",
                        "in": "path",
                        "description": "house id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "upload item image success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "links": {
                            "uploadImage": {
                                "operationId": "storeImage",
                                "parameters": {
                                    "image": "public/images/3.jpg"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "App_Key": "",
                        "Bearer_auth": "",
                        "Provider": ""
                    }
                ]
            }
        },
        "/houses/{house_id}/available": {
            "put": {
                "tags": [
                    "Houses"
                ],
                "summary": "mark house as available",
                "description": "mark house as available",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\House\\HouseController::markAvailable",
                "parameters": [
                    {
                        "name": "house_id",
                        "in": "path",
                        "description": "house id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Add PUT in the field to update",
                                        "type": "string",
                                        "enum": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "update house success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "links": {
                            "markAvailable": {
                                "operationId": "markAvailable",
                                "parameters": {
                                    "_method": "PUT"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "App_Key": "",
                        "Bearer_auth": ""
                    }
                ]
            }
        },
        "/houses/{house_id}/unavailable": {
            "put": {
                "tags": [
                    "Houses"
                ],
                "summary": "mark house as unavailable",
                "description": "mark house as unavailable",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\House\\HouseController::markUnavailable",
                "parameters": [
                    {
                        "name": "house_id",
                        "in": "path",
                        "description": "house id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Add PUT in the field to update",
                                        "type": "string",
                                        "enum": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "update house success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "links": {
                            "markUnavailable": {
                                "operationId": "markUnavailable",
                                "parameters": {
                                    "_method": "PUT"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "App_Key": "",
                        "Bearer_auth": ""
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer_auth": {
                "type": "http",
                "scheme": "bearer"
            },
            "App_Key": {
                "type": "apiKey",
                "name": "X-APP-TOKEN",
                "in": "header"
            }
        }
    }
}